pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.0.3"
mysqlclient = "^2.2.4"
pymysql = "^1.1.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
pyjwt = "^2.8.0"
drf-yasg = "^1.21.7"
setuptools = "^69.2.0"
python-decouple = "^3.8"
django-environ = "^0.11.2"
pytest-django = "^4.8.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
